convert costs from dollars to cents

+++ db/migrate/20140904174216_add_cost_fees_and_total_cost_to_subscriptions.rb
@@ -1,7 +1,7 @@
 class AddCostFeesAndTotalCostToSubscriptions < ActiveRecord::Migration
   def change
-    add_column :subscriptions, :cost, :float
-    add_column :subscriptions, :fees, :float
-    add_column :subscriptions, :total_cost, :float
+    add_column :subscriptions, :cost, :integer
+    add_column :subscriptions, :fees, :integer
+    add_column :subscriptions, :total_cost, :integer
   end
 end
